flask/requirements.txt
json
Flask
psycopg2-binary
sqlachemy
python-dotenv
